+++ lib/platform/times/platform.times.mapper.ts
@@ -19,6 +19,9 @@ export function mapToPlatformTimes(
     alerts: propertyToArray(platform.Alert).map(mapToAlert),
     content: mapToContent(positionJson[API_TIMES].Content),
     name: platform.$.Name,
+    number: platform.$.PlatformNo
+      ? parseInt(platform.$.PlatformNo, 10)
+      : undefined,
     routes: propertyToArray(platform.Route).map(mapToRoute),
     tag: parseInt(platform.$.PlatformTag, 10),
   };
@@ -41,8 +44,10 @@ function mapToRoute(routeJson: ElementCompact): Route {
 
 function mapToTrip(tripJson: ElementCompact): Trip {
   return {
-    eta: tripJson.$.ETA,
-    id: tripJson.$.TripNo,
-    wheelchairAccess: tripJson.$.WheelchairAccess,
+    eta: parseInt(tripJson.$.ETA, 10),
+    id: tripJson.$.TripID ? parseInt(tripJson.$.TripID, 10) : undefined,
+    wheelchairAccess: tripJson.$.WheelchairAccess
+      ? tripJson.$.WheelchairAccess === 'true'
+      : undefined,
   };
 }